+++ DatadogCore/Tests/Datadog/Tracing/TracingURLSessionHandlerTests.swift
@@ -57,7 +57,7 @@ class TracingURLSessionHandlerTests: XCTestCase {
 
         // Given
         let request: ImmutableRequest = .mockWith(httpMethod: "POST")
-        let interception = URLSessionTaskInterception(request: request, isFirstParty: true, trackingMode: .metrics)
+        let interception = URLSessionTaskInterception(request: request, isFirstParty: true, trackingMode: .registeredDelegate)
         interception.register(metrics: .mockAny())
         interception.register(response: .mockResponseWith(statusCode: 200), error: nil)
 
@@ -85,7 +85,7 @@ class TracingURLSessionHandlerTests: XCTestCase {
             httpMethod: "GET"
         )
         let error = NSError(domain: "domain", code: 123, userInfo: [NSLocalizedDescriptionKey: "network error"])
-        let interception = URLSessionTaskInterception(request: request, isFirstParty: true, trackingMode: .metrics)
+        let interception = URLSessionTaskInterception(request: request, isFirstParty: true, trackingMode: .registeredDelegate)
         interception.register(response: nil, error: error)
         interception.register(
             metrics: .mockWith(
@@ -154,7 +154,7 @@ class TracingURLSessionHandlerTests: XCTestCase {
 
         // Given
         let request: ImmutableRequest = .mockWith(httpMethod: "GET")
-        let interception = URLSessionTaskInterception(request: request, isFirstParty: true, trackingMode: .metrics)
+        let interception = URLSessionTaskInterception(request: request, isFirstParty: true, trackingMode: .registeredDelegate)
         interception.register(response: .mockResponseWith(statusCode: 404), error: nil)
         interception.register(
             metrics: .mockWith(